requirements.txt
black
pytest
ruff

flask
flask-cors
python-dotenv
openai